pyproject.toml
[project]
name = "article-saver"
version = "0.1.0"
description = "Helps me save articles for later reading."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "feedgen>=1.0.0",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.37.0",
    "trafilatura>=1.12.2",
    "nh3>=0.2.18",
    "sqlalchemy>=2.0.36",
    "psycopg2-binary>=2.9.10",
    "httpx>=0.28.1",
]
